Email network (directed)

A list of two datasets, vertices and edges, containing data on employees and (fictitious) email headers from two weeks of internal GAStech company email between employees made available as part of the VAST challenge 2014 (K. Cook, G. Grinstein, M. Whiting, see http://www.cs.umd.edu/hcil/varepository/benchmarks.php). The variables are as follows:

Usage

email

Format

A list of two data frames:

the edges data set consists of ten variables of length 9063:

From: Email address of sender

eID: email ID. If an email ws sent to multiple recipients, the email ID is the same.

Date: Date and time email was sent

Subject: Subject line of email

to: Email address of recipient

month: Month email was sent

day: Day of month email was sent

year: Year email was sent

nrecipients: Number of recipients of email

the nodes data set consists of 18 variables with information on 55 employees:

label: Employee's email address

LastName: Employee's last name

FirstName: Employee's first name

BirthDate: Employee's first date

BirthCountry: Employee's (fictional) country of birth

Gender: Employee's gender

CitizenshipCountry: Employee's (fictional) country of citizenship

CitizenshipBasis: Is citizenship by birth or parents?

CitizenshipStartDate: When employee became a citizen

PassportCountry: (Fictional) country issuing employee's passport

PassportIssueDate: Date employee received passport

PassportExpirationDate: Date employee's passport expires

CurrentEmploymentType: Employee's department in the company

CurrentEmploymentTitle: Employee's title at the company

CurrentEmploymentStartDate: Date employee started at position

MilitaryServiceBranch: Branch of the (fictional) military in which the employee serves

MilitaryDischargeType: General or honorable discharge from military service?