Function for converting a network edge list in data frame form into the correct format for use with geomnet

Function for converting a network edge list in data frame form into the correct format for use with geomnet

Usage

# S3 method for class 'edgedf'
fortify(model, data, group = NULL, ...)

Arguments

model

A network edgelist of class "edgedf". See as.edgedf. Can contain edge variables as well.

data

Data frame containing network node list and other node information. First column should contain node ids.

group

character. Used for facetting. If you wish to facet on network variable provide the name of that variable here.

...

not used in this function

Examples

data(blood)
fortify(as.edgedf(blood$edges), blood$vertices)
#> Using from as the from node column and to as the to node column.
#> If this is not correct, rewrite dat so that the first 2 columns are from and to node, respectively.
#> Joining edge and node information by from_id and label respectively.
#>     from_id to_id group_to type rho        Ethnicity Predominance
#> 1        A+   AB+     diff    A pos       Caucasians         33.0
#> 2        A+   AB+     diff    A pos            Asian         27.0
#> 3        A+   AB+     diff    A pos         Hispanic         29.0
#> 4        A+   AB+     diff    A pos African.American         24.0
#> 5        A+    A+     same    A pos       Caucasians         33.0
#> 6        A+    A+     same    A pos            Asian         27.0
#> 7        A+    A+     same    A pos         Hispanic         29.0
#> 8        A+    A+     same    A pos African.American         24.0
#> 9        A-   AB+     diff    A neg            Asian          0.5
#> 10       A-   AB+     diff    A neg         Hispanic          2.0
#> 11       A-   AB+     diff    A neg African.American          2.0
#> 12       A-   AB+     diff    A neg       Caucasians          7.0
#> 13       A-   AB-     diff    A neg            Asian          0.5
#> 14       A-   AB-     diff    A neg         Hispanic          2.0
#> 15       A-   AB-     diff    A neg African.American          2.0
#> 16       A-   AB-     diff    A neg       Caucasians          7.0
#> 17       A-    A+     same    A neg            Asian          0.5
#> 18       A-    A+     same    A neg         Hispanic          2.0
#> 19       A-    A+     same    A neg African.American          2.0
#> 20       A-    A+     same    A neg       Caucasians          7.0
#> 21       A-    A-     same    A neg            Asian          0.5
#> 22       A-    A-     same    A neg         Hispanic          2.0
#> 23       A-    A-     same    A neg African.American          2.0
#> 24       A-    A-     same    A neg       Caucasians          7.0
#> 25      AB+   AB+     same   AB pos African.American          4.0
#> 26      AB+   AB+     same   AB pos       Caucasians          3.0
#> 27      AB+   AB+     same   AB pos         Hispanic          2.0
#> 28      AB+   AB+     same   AB pos            Asian          7.0
#> 29      AB-   AB+     same   AB neg African.American          0.3
#> 30      AB-   AB+     same   AB neg            Asian          0.1
#> 31      AB-   AB+     same   AB neg       Caucasians          1.0
#> 32      AB-   AB+     same   AB neg         Hispanic          0.2
#> 33      AB-   AB-     same   AB neg African.American          0.3
#> 34      AB-   AB-     same   AB neg            Asian          0.1
#> 35      AB-   AB-     same   AB neg       Caucasians          1.0
#> 36      AB-   AB-     same   AB neg         Hispanic          0.2
#> 37       B+   AB+     diff    B pos            Asian         25.0
#> 38       B+   AB+     diff    B pos         Hispanic          9.0
#> 39       B+   AB+     diff    B pos       Caucasians          9.0
#> 40       B+   AB+     diff    B pos African.American         18.0
#> 41       B+    B+     same    B pos            Asian         25.0
#> 42       B+    B+     same    B pos         Hispanic          9.0
#> 43       B+    B+     same    B pos       Caucasians          9.0
#> 44       B+    B+     same    B pos African.American         18.0
#> 45       B-   AB+     diff    B neg         Hispanic          1.0
#> 46       B-   AB+     diff    B neg            Asian          0.4
#> 47       B-   AB+     diff    B neg African.American          1.0
#> 48       B-   AB+     diff    B neg       Caucasians          2.0
#> 49       B-   AB-     diff    B neg         Hispanic          1.0
#> 50       B-   AB-     diff    B neg            Asian          0.4
#> 51       B-   AB-     diff    B neg African.American          1.0
#> 52       B-   AB-     diff    B neg       Caucasians          2.0
#> 53       B-    B+     same    B neg         Hispanic          1.0
#> 54       B-    B+     same    B neg            Asian          0.4
#> 55       B-    B+     same    B neg African.American          1.0
#> 56       B-    B+     same    B neg       Caucasians          2.0
#> 57       B-    B-     same    B neg         Hispanic          1.0
#> 58       B-    B-     same    B neg            Asian          0.4
#> 59       B-    B-     same    B neg African.American          1.0
#> 60       B-    B-     same    B neg       Caucasians          2.0
#> 61       O+   AB+     diff    O pos            Asian         39.0
#> 62       O+   AB+     diff    O pos         Hispanic         53.0
#> 63       O+   AB+     diff    O pos       Caucasians         37.0
#> 64       O+   AB+     diff    O pos African.American         47.0
#> 65       O+    A+     diff    O pos            Asian         39.0
#> 66       O+    A+     diff    O pos         Hispanic         53.0
#> 67       O+    A+     diff    O pos       Caucasians         37.0
#> 68       O+    A+     diff    O pos African.American         47.0
#> 69       O+    B+     diff    O pos            Asian         39.0
#> 70       O+    B+     diff    O pos         Hispanic         53.0
#> 71       O+    B+     diff    O pos       Caucasians         37.0
#> 72       O+    B+     diff    O pos African.American         47.0
#> 73       O+    O+     same    O pos            Asian         39.0
#> 74       O+    O+     same    O pos         Hispanic         53.0
#> 75       O+    O+     same    O pos       Caucasians         37.0
#> 76       O+    O+     same    O pos African.American         47.0
#> 77       O-    A+     diff    O neg African.American          4.0
#> 78       O-    A+     diff    O neg            Asian          1.0
#> 79       O-    A+     diff    O neg       Caucasians          8.0
#> 80       O-    A+     diff    O neg         Hispanic          4.0
#> 81       O-    A-     diff    O neg African.American          4.0
#> 82       O-    A-     diff    O neg            Asian          1.0
#> 83       O-    A-     diff    O neg       Caucasians          8.0
#> 84       O-    A-     diff    O neg         Hispanic          4.0
#> 85       O-   AB+     diff    O neg African.American          4.0
#> 86       O-   AB+     diff    O neg            Asian          1.0
#> 87       O-   AB+     diff    O neg       Caucasians          8.0
#> 88       O-   AB+     diff    O neg         Hispanic          4.0
#> 89       O-   AB-     diff    O neg African.American          4.0
#> 90       O-   AB-     diff    O neg            Asian          1.0
#> 91       O-   AB-     diff    O neg       Caucasians          8.0
#> 92       O-   AB-     diff    O neg         Hispanic          4.0
#> 93       O-    B+     diff    O neg African.American          4.0
#> 94       O-    B+     diff    O neg            Asian          1.0
#> 95       O-    B+     diff    O neg       Caucasians          8.0
#> 96       O-    B+     diff    O neg         Hispanic          4.0
#> 97       O-    B-     diff    O neg African.American          4.0
#> 98       O-    B-     diff    O neg            Asian          1.0
#> 99       O-    B-     diff    O neg       Caucasians          8.0
#> 100      O-    B-     diff    O neg         Hispanic          4.0
#> 101      O-    O+     same    O neg African.American          4.0
#> 102      O-    O+     same    O neg            Asian          1.0
#> 103      O-    O+     same    O neg       Caucasians          8.0
#> 104      O-    O+     same    O neg         Hispanic          4.0
#> 105      O-    O-     same    O neg African.American          4.0
#> 106      O-    O-     same    O neg            Asian          1.0
#> 107      O-    O-     same    O neg       Caucasians          8.0
#> 108      O-    O-     same    O neg         Hispanic          4.0
fortify(as.edgedf(blood$edges), blood$vertices, group = "Ethnicity")
#> Using from as the from node column and to as the to node column.
#> If this is not correct, rewrite dat so that the first 2 columns are from and to node, respectively.
#> Joining edge and node information by from_id and label respectively.
#>     from to_id group_to        Ethnicity type rho Predominance
#> 1     A+   AB+     diff African.American    A pos         24.0
#> 2     A+   AB+     diff            Asian    A pos         27.0
#> 3     A+   AB+     diff       Caucasians    A pos         33.0
#> 4     A+   AB+     diff         Hispanic    A pos         29.0
#> 5     A+    A+     same African.American    A pos         24.0
#> 6     A+    A+     same            Asian    A pos         27.0
#> 7     A+    A+     same       Caucasians    A pos         33.0
#> 8     A+    A+     same         Hispanic    A pos         29.0
#> 9     A-   AB+     diff African.American    A neg          2.0
#> 10    A-   AB+     diff            Asian    A neg          0.5
#> 11    A-   AB+     diff       Caucasians    A neg          7.0
#> 12    A-   AB+     diff         Hispanic    A neg          2.0
#> 13    A-   AB-     diff African.American    A neg          2.0
#> 14    A-   AB-     diff            Asian    A neg          0.5
#> 15    A-   AB-     diff       Caucasians    A neg          7.0
#> 16    A-   AB-     diff         Hispanic    A neg          2.0
#> 17    A-    A+     same African.American    A neg          2.0
#> 18    A-    A+     same            Asian    A neg          0.5
#> 19    A-    A+     same       Caucasians    A neg          7.0
#> 20    A-    A+     same         Hispanic    A neg          2.0
#> 21    A-    A-     same African.American    A neg          2.0
#> 22    A-    A-     same            Asian    A neg          0.5
#> 23    A-    A-     same       Caucasians    A neg          7.0
#> 24    A-    A-     same         Hispanic    A neg          2.0
#> 25   AB+   AB+     same African.American   AB pos          4.0
#> 26   AB+   AB+     same            Asian   AB pos          7.0
#> 27   AB+   AB+     same       Caucasians   AB pos          3.0
#> 28   AB+   AB+     same         Hispanic   AB pos          2.0
#> 29   AB-   AB+     same African.American   AB neg          0.3
#> 30   AB-   AB+     same            Asian   AB neg          0.1
#> 31   AB-   AB+     same       Caucasians   AB neg          1.0
#> 32   AB-   AB+     same         Hispanic   AB neg          0.2
#> 33   AB-   AB-     same African.American   AB neg          0.3
#> 34   AB-   AB-     same            Asian   AB neg          0.1
#> 35   AB-   AB-     same       Caucasians   AB neg          1.0
#> 36   AB-   AB-     same         Hispanic   AB neg          0.2
#> 37    B+   AB+     diff African.American    B pos         18.0
#> 38    B+   AB+     diff            Asian    B pos         25.0
#> 39    B+   AB+     diff       Caucasians    B pos          9.0
#> 40    B+   AB+     diff         Hispanic    B pos          9.0
#> 41    B+    B+     same African.American    B pos         18.0
#> 42    B+    B+     same            Asian    B pos         25.0
#> 43    B+    B+     same       Caucasians    B pos          9.0
#> 44    B+    B+     same         Hispanic    B pos          9.0
#> 45    B-   AB+     diff African.American    B neg          1.0
#> 46    B-   AB+     diff            Asian    B neg          0.4
#> 47    B-   AB+     diff       Caucasians    B neg          2.0
#> 48    B-   AB+     diff         Hispanic    B neg          1.0
#> 49    B-   AB-     diff African.American    B neg          1.0
#> 50    B-   AB-     diff            Asian    B neg          0.4
#> 51    B-   AB-     diff       Caucasians    B neg          2.0
#> 52    B-   AB-     diff         Hispanic    B neg          1.0
#> 53    B-    B+     same African.American    B neg          1.0
#> 54    B-    B+     same            Asian    B neg          0.4
#> 55    B-    B+     same       Caucasians    B neg          2.0
#> 56    B-    B+     same         Hispanic    B neg          1.0
#> 57    B-    B-     same African.American    B neg          1.0
#> 58    B-    B-     same            Asian    B neg          0.4
#> 59    B-    B-     same       Caucasians    B neg          2.0
#> 60    B-    B-     same         Hispanic    B neg          1.0
#> 61    O+   AB+     diff African.American    O pos         47.0
#> 62    O+   AB+     diff            Asian    O pos         39.0
#> 63    O+   AB+     diff       Caucasians    O pos         37.0
#> 64    O+   AB+     diff         Hispanic    O pos         53.0
#> 65    O+    A+     diff African.American    O pos         47.0
#> 66    O+    A+     diff            Asian    O pos         39.0
#> 67    O+    A+     diff       Caucasians    O pos         37.0
#> 68    O+    A+     diff         Hispanic    O pos         53.0
#> 69    O+    B+     diff African.American    O pos         47.0
#> 70    O+    B+     diff            Asian    O pos         39.0
#> 71    O+    B+     diff       Caucasians    O pos         37.0
#> 72    O+    B+     diff         Hispanic    O pos         53.0
#> 73    O+    O+     same African.American    O pos         47.0
#> 74    O+    O+     same            Asian    O pos         39.0
#> 75    O+    O+     same       Caucasians    O pos         37.0
#> 76    O+    O+     same         Hispanic    O pos         53.0
#> 77    O-    A+     diff African.American    O neg          4.0
#> 78    O-    A+     diff            Asian    O neg          1.0
#> 79    O-    A+     diff       Caucasians    O neg          8.0
#> 80    O-    A+     diff         Hispanic    O neg          4.0
#> 81    O-    A-     diff African.American    O neg          4.0
#> 82    O-    A-     diff            Asian    O neg          1.0
#> 83    O-    A-     diff       Caucasians    O neg          8.0
#> 84    O-    A-     diff         Hispanic    O neg          4.0
#> 85    O-   AB+     diff African.American    O neg          4.0
#> 86    O-   AB+     diff            Asian    O neg          1.0
#> 87    O-   AB+     diff       Caucasians    O neg          8.0
#> 88    O-   AB+     diff         Hispanic    O neg          4.0
#> 89    O-   AB-     diff African.American    O neg          4.0
#> 90    O-   AB-     diff            Asian    O neg          1.0
#> 91    O-   AB-     diff       Caucasians    O neg          8.0
#> 92    O-   AB-     diff         Hispanic    O neg          4.0
#> 93    O-    B+     diff African.American    O neg          4.0
#> 94    O-    B+     diff            Asian    O neg          1.0
#> 95    O-    B+     diff       Caucasians    O neg          8.0
#> 96    O-    B+     diff         Hispanic    O neg          4.0
#> 97    O-    B-     diff African.American    O neg          4.0
#> 98    O-    B-     diff            Asian    O neg          1.0
#> 99    O-    B-     diff       Caucasians    O neg          8.0
#> 100   O-    B-     diff         Hispanic    O neg          4.0
#> 101   O-    O+     same African.American    O neg          4.0
#> 102   O-    O+     same            Asian    O neg          1.0
#> 103   O-    O+     same       Caucasians    O neg          8.0
#> 104   O-    O+     same         Hispanic    O neg          4.0
#> 105   O-    O-     same African.American    O neg          4.0
#> 106   O-    O-     same            Asian    O neg          1.0
#> 107   O-    O-     same       Caucasians    O neg          8.0
#> 108   O-    O-     same         Hispanic    O neg          4.0