Function for converting an igraph object into the correct format for use with geomnet

Function for converting an igraph object into the correct format for use with geomnet

Usage

# S3 method for class 'igraph'
fortify(model, data = NULL, group = NULL, ...)

Arguments

model

A network object of class "igraph".

data

NULL - not used in this function

group

character. Used for facetting. If you wish to facet on a network variable provide the name of that variable here.

...

not used in this function

Examples

# class igraph (igraph, igraphdata packages)
library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     as_data_frame, groups, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
data(blood, package = "geomnet")
blood.igraph <- graph_from_data_frame(d = blood$edges, 
   directed = TRUE, vertices = unique(blood$vertices[, 1:3]))
fortify(blood.igraph)
#>    from  to group_to type rho
#> 1    A+ AB+     diff    A pos
#> 2    A+  A+     same    A pos
#> 3    A- AB+     diff    A neg
#> 4    A- AB-     diff    A neg
#> 5    A-  A+     same    A neg
#> 6    A-  A-     same    A neg
#> 7   AB+ AB+     same   AB pos
#> 8   AB- AB+     same   AB neg
#> 9   AB- AB-     same   AB neg
#> 10   B+ AB+     diff    B pos
#> 11   B+  B+     same    B pos
#> 12   B- AB+     diff    B neg
#> 13   B- AB-     diff    B neg
#> 14   B-  B+     same    B neg
#> 15   B-  B-     same    B neg
#> 16   O+ AB+     diff    O pos
#> 17   O+  A+     diff    O pos
#> 18   O+  B+     diff    O pos
#> 19   O+  O+     same    O pos
#> 20   O-  A+     diff    O neg
#> 21   O-  A-     diff    O neg
#> 22   O- AB+     diff    O neg
#> 23   O- AB-     diff    O neg
#> 24   O-  B+     diff    O neg
#> 25   O-  B-     diff    O neg
#> 26   O-  O+     same    O neg
#> 27   O-  O-     same    O neg