Function for converting a network object into the correct format for use with geomnet

Function for converting a network object into the correct format for use with geomnet

Usage

# S3 method for class 'network'
fortify(model, data = NULL, group = NULL, ...)

Arguments

model

object of class "network"

data

NULL - not used in this function

group

character. Used for facetting. If you wish to facet on a network variable provide the name of that variable here.

...

not used in this function

Examples

# class network (sna, network, statnet packages)

library(network)
#> 
#> ‘network’ 1.18.2 (2023-12-04), part of the Statnet Project
#> * ‘news(package="network")’ for changes since last version
#> * ‘citation("network")’ for citation information
#> * ‘https://statnet.org’ for help, support, and other information
#> 
#> Attaching package: ‘network’
#> The following objects are masked from ‘package:igraph’:
#> 
#>     %c%, %s%, add.edges, add.vertices, delete.edges, delete.vertices,
#>     get.edge.attribute, get.edges, get.vertex.attribute, is.bipartite,
#>     is.directed, list.edge.attributes, list.vertex.attributes,
#>     set.edge.attribute, set.vertex.attribute
data(emon, package = "network")
fortify(emon$Cheyenne)
#>    from to Frequency  na.x Command.Rank.Score Decision.Rank.Score Formalization
#> 1     1  2         1 FALSE                  0                  20             2
#> 2     1  3         1 FALSE                  0                  20             2
#> 3     1  4         1 FALSE                  0                  20             2
#> 4     1  5         1 FALSE                  0                  20             2
#> 5     1  7         1 FALSE                  0                  20             2
#> 6     1  8         1 FALSE                  0                  20             2
#> 7     1  9         3 FALSE                  0                  20             2
#> 8     1 10         1 FALSE                  0                  20             2
#> 9     1 11         1 FALSE                  0                  20             2
#> 10    1 14         1 FALSE                  0                  20             2
#> 11    2  1         1 FALSE                 10                   7             1
#> 12    2  3         3 FALSE                 10                   7             1
#> 13    2  7         1 FALSE                 10                   7             1
#> 14    2  8         4 FALSE                 10                   7             1
#> 15    2  9         2 FALSE                 10                   7             1
#> 16    2 10         2 FALSE                 10                   7             1
#> 17    2 14         3 FALSE                 10                   7             1
#> 18    3  1         2 FALSE                  3                   0             1
#> 19    3  2         1 FALSE                  3                   0             1
#> 20    3  4         4 FALSE                  3                   0             1
#> 21    3  5         3 FALSE                  3                   0             1
#> 22    3  6         3 FALSE                  3                   0             1
#> 23    3  7         4 FALSE                  3                   0             1
#> 24    3  8         3 FALSE                  3                   0             1
#> 25    3  9         1 FALSE                  3                   0             1
#> 26    3 10         1 FALSE                  3                   0             1
#> 27    3 11         1 FALSE                  3                   0             1
#> 28    3 12         1 FALSE                  3                   0             1
#> 29    3 13         3 FALSE                  3                   0             1
#> 30    4  2         4 FALSE                  5                   5             1
#> 31    4  3         2 FALSE                  5                   5             1
#> 32    4  7         1 FALSE                  5                   5             1
#> 33    4  9         1 FALSE                  5                   5             1
#> 34    4 10         1 FALSE                  5                   5             1
#> 35    5  7         4 FALSE                  0                   0             1
#> 36    5  8         1 FALSE                  0                   0             1
#> 37    5 13         1 FALSE                  0                   0             1
#> 38    6 NA        NA    NA                  0                   0             1
#> 39    7  3         1 FALSE                 20                  20             1
#> 40    7  8         1 FALSE                 20                  20             1
#> 41    7 10         1 FALSE                 20                  20             1
#> 42    7 11         2 FALSE                 20                  20             1
#> 43    7 12         1 FALSE                 20                  20             1
#> 44    7 13         1 FALSE                 20                  20             1
#> 45    8  1         1 FALSE                 40                  50             2
#> 46    8  2         1 FALSE                 40                  50             2
#> 47    8  4         1 FALSE                 40                  50             2
#> 48    8  5         1 FALSE                 40                  50             2
#> 49    8  6         3 FALSE                 40                  50             2
#> 50    8  7         3 FALSE                 40                  50             2
#> 51    8  9         1 FALSE                 40                  50             2
#> 52    8 10         1 FALSE                 40                  50             2
#> 53    8 11         1 FALSE                 40                  50             2
#> 54    8 12         4 FALSE                 40                  50             2
#> 55    8 13         1 FALSE                 40                  50             2
#> 56    8 14         1 FALSE                 40                  50             2
#> 57    9  1         1 FALSE                 10                  10             1
#> 58    9  6         1 FALSE                 10                  10             1
#> 59    9  7         3 FALSE                 10                  10             1
#> 60    9  8         1 FALSE                 10                  10             1
#> 61    9 10         1 FALSE                 10                  10             1
#> 62    9 13         1 FALSE                 10                  10             1
#> 63    9 14         1 FALSE                 10                  10             1
#> 64   10  7         1 FALSE                 30                  20             3
#> 65   10  9         1 FALSE                 30                  20             3
#> 66   11  7         2 FALSE                  2                   3            NA
#> 67   11  9         2 FALSE                  2                   3            NA
#> 68   11 10         3 FALSE                  2                   3            NA
#> 69   11 13         1 FALSE                  2                   3            NA
#> 70   11 14         2 FALSE                  2                   3            NA
#> 71   12  3         2 FALSE                  2                   3             1
#> 72   12  7         3 FALSE                  2                   3             1
#> 73   12  8         1 FALSE                  2                   3             1
#> 74   12  9         1 FALSE                  2                   3             1
#> 75   13  3         4 FALSE                  2                   2             1
#> 76   13  7         3 FALSE                  2                   2             1
#> 77   13  8         3 FALSE                  2                   2             1
#> 78   13  9         1 FALSE                  2                   2             1
#> 79   13 10         3 FALSE                  2                   2             1
#> 80   13 11         1 FALSE                  2                   2             1
#> 81   13 12         1 FALSE                  2                   2             1
#> 82   14  1         1 FALSE                  0                   0             1
#> 83   14  5         1 FALSE                  0                   0             1
#> 84   14  8         1 FALSE                  0                   0             1
#>    Location Paid.Staff Sponsorship Volunteer.Staff  na.y
#> 1         L         10       State              50 FALSE
#> 2         L         10       State              50 FALSE
#> 3         L         10       State              50 FALSE
#> 4         L         10       State              50 FALSE
#> 5         L         10       State              50 FALSE
#> 6         L         10       State              50 FALSE
#> 7         L         10       State              50 FALSE
#> 8         L         10       State              50 FALSE
#> 9         L         10       State              50 FALSE
#> 10        L         10       State              50 FALSE
#> 11        L        400       State            2000 FALSE
#> 12        L        400       State            2000 FALSE
#> 13        L        400       State            2000 FALSE
#> 14        L        400       State            2000 FALSE
#> 15        L        400       State            2000 FALSE
#> 16        L        400       State            2000 FALSE
#> 17        L        400       State            2000 FALSE
#> 18        L        200       State               0 FALSE
#> 19        L        200       State               0 FALSE
#> 20        L        200       State               0 FALSE
#> 21        L        200       State               0 FALSE
#> 22        L        200       State               0 FALSE
#> 23        L        200       State               0 FALSE
#> 24        L        200       State               0 FALSE
#> 25        L        200       State               0 FALSE
#> 26        L        200       State               0 FALSE
#> 27        L        200       State               0 FALSE
#> 28        L        200       State               0 FALSE
#> 29        L        200       State               0 FALSE
#> 30        L         60     Federal               0 FALSE
#> 31        L         60     Federal               0 FALSE
#> 32        L         60     Federal               0 FALSE
#> 33        L         60     Federal               0 FALSE
#> 34        L         60     Federal               0 FALSE
#> 35        L          1     Private              20 FALSE
#> 36        L          1     Private              20 FALSE
#> 37        L          1     Private              20 FALSE
#> 38        L          7     Private              80 FALSE
#> 39        L         60      County              20 FALSE
#> 40        L         60      County              20 FALSE
#> 41        L         60      County              20 FALSE
#> 42        L         60      County              20 FALSE
#> 43        L         60      County              20 FALSE
#> 44        L         60      County              20 FALSE
#> 45        L          7 County/City             100 FALSE
#> 46        L          7 County/City             100 FALSE
#> 47        L          7 County/City             100 FALSE
#> 48        L          7 County/City             100 FALSE
#> 49        L          7 County/City             100 FALSE
#> 50        L          7 County/City             100 FALSE
#> 51        L          7 County/City             100 FALSE
#> 52        L          7 County/City             100 FALSE
#> 53        L          7 County/City             100 FALSE
#> 54        L          7 County/City             100 FALSE
#> 55        L          7 County/City             100 FALSE
#> 56        L          7 County/City             100 FALSE
#> 57        L         70        City               0 FALSE
#> 58        L         70        City               0 FALSE
#> 59        L         70        City               0 FALSE
#> 60        L         70        City               0 FALSE
#> 61        L         70        City               0 FALSE
#> 62        L         70        City               0 FALSE
#> 63        L         70        City               0 FALSE
#> 64        L        100        City               0 FALSE
#> 65        L        100        City               0 FALSE
#> 66        L         NA      County              NA FALSE
#> 67        L         NA      County              NA FALSE
#> 68        L         NA      County              NA FALSE
#> 69        L         NA      County              NA FALSE
#> 70        L         NA      County              NA FALSE
#> 71        L          1      County              20 FALSE
#> 72        L          1      County              20 FALSE
#> 73        L          1      County              20 FALSE
#> 74        L          1      County              20 FALSE
#> 75        L         10     Private               6 FALSE
#> 76        L         10     Private               6 FALSE
#> 77        L         10     Private               6 FALSE
#> 78        L         10     Private               6 FALSE
#> 79        L         10     Private               6 FALSE
#> 80        L         10     Private               6 FALSE
#> 81        L         10     Private               6 FALSE
#> 82        L          0     Private              NA FALSE
#> 83        L          0     Private              NA FALSE
#> 84        L          0     Private              NA FALSE
#>                                      vertex.names
#> 1       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 2       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 3       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 4       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 5       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 6       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 7       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 8       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 9       Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 10      Wyoming.Disaster.and.Civil.Defense.Agnecy
#> 11    Wyoming.State.National.Guard..Army.and.Air.
#> 12    Wyoming.State.National.Guard..Army.and.Air.
#> 13    Wyoming.State.National.Guard..Army.and.Air.
#> 14    Wyoming.State.National.Guard..Army.and.Air.
#> 15    Wyoming.State.National.Guard..Army.and.Air.
#> 16    Wyoming.State.National.Guard..Army.and.Air.
#> 17    Wyoming.State.National.Guard..Army.and.Air.
#> 18                   Wyoming.State.Highway.Patrol
#> 19                   Wyoming.State.Highway.Patrol
#> 20                   Wyoming.State.Highway.Patrol
#> 21                   Wyoming.State.Highway.Patrol
#> 22                   Wyoming.State.Highway.Patrol
#> 23                   Wyoming.State.Highway.Patrol
#> 24                   Wyoming.State.Highway.Patrol
#> 25                   Wyoming.State.Highway.Patrol
#> 26                   Wyoming.State.Highway.Patrol
#> 27                   Wyoming.State.Highway.Patrol
#> 28                   Wyoming.State.Highway.Patrol
#> 29                   Wyoming.State.Highway.Patrol
#> 30   Francis.E..Warren.Air.Force.Base..fire.unit.
#> 31   Francis.E..Warren.Air.Force.Base..fire.unit.
#> 32   Francis.E..Warren.Air.Force.Base..fire.unit.
#> 33   Francis.E..Warren.Air.Force.Base..fire.unit.
#> 34   Francis.E..Warren.Air.Force.Base..fire.unit.
#> 35                                      Red.Cross
#> 36                                      Red.Cross
#> 37                                      Red.Cross
#> 38                                 Salvation.Army
#> 39                Laramie.County.Sheriff.s.Office
#> 40                Laramie.County.Sheriff.s.Office
#> 41                Laramie.County.Sheriff.s.Office
#> 42                Laramie.County.Sheriff.s.Office
#> 43                Laramie.County.Sheriff.s.Office
#> 44                Laramie.County.Sheriff.s.Office
#> 45 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 46 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 47 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 48 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 49 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 50 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 51 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 52 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 53 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 54 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 55 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 56 Laramie.County...Cheyenne.Civil.Defense.Agency
#> 57                       Cheyenne.Fire.Department
#> 58                       Cheyenne.Fire.Department
#> 59                       Cheyenne.Fire.Department
#> 60                       Cheyenne.Fire.Department
#> 61                       Cheyenne.Fire.Department
#> 62                       Cheyenne.Fire.Department
#> 63                       Cheyenne.Fire.Department
#> 64                     Cheyenne.Police.Department
#> 65                     Cheyenne.Police.Department
#> 66    Laramie.County.Fire.District..1..volunteer.
#> 67    Laramie.County.Fire.District..1..volunteer.
#> 68    Laramie.County.Fire.District..1..volunteer.
#> 69    Laramie.County.Fire.District..1..volunteer.
#> 70    Laramie.County.Fire.District..1..volunteer.
#> 71    Laramie.County.Fire.District..2..volunteer.
#> 72    Laramie.County.Fire.District..2..volunteer.
#> 73    Laramie.County.Fire.District..2..volunteer.
#> 74    Laramie.County.Fire.District..2..volunteer.
#> 75                          A.1.Ambulance.Service
#> 76                          A.1.Ambulance.Service
#> 77                          A.1.Ambulance.Service
#> 78                          A.1.Ambulance.Service
#> 79                          A.1.Ambulance.Service
#> 80                          A.1.Ambulance.Service
#> 81                          A.1.Ambulance.Service
#> 82                          Shy.Wy.HAM.Radio.Club
#> 83                          Shy.Wy.HAM.Radio.Club
#> 84                          Shy.Wy.HAM.Radio.Club