Function for converting a network adjacency matrix into the correct format for use with geomnet

Function for converting a network adjacency matrix into the correct format for use with geomnet

# S3 method for adjmat
fortify(model, data = NULL, ...)

Arguments

model
An adjacency matrix of class "adjmat".

data
not used in this function

...
not used in this function

Examples

data(emon, package = "network")
adjmat <- as.adjmat(network::as.matrix.network.adjacency(emon$MtSi))
str(adjmat)

#>  'adjmat' num [1:13, 1:13] 0 0 1 0 1 1 0 1 0 1 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:13] "King.County.Police" "Skagit.County.Sheriff.s.Office" "Seattle.ESAR..Explorer.Search.and.Rescue." "Tacoma.Pierce.County.ESAR.Unit" ...
#>   ..$ : chr [1:13] "King.County.Police" "Skagit.County.Sheriff.s.Office" "Seattle.ESAR..Explorer.Search.and.Rescue." "Tacoma.Pierce.County.ESAR.Unit" ...

fortify(adjmat)

#>                                                        from
#> 1            Boeing.Employees.Amateur.Radio.Society..BEARS.
#> 2                                    Boeing.Helicopter.Team
#> 3                                    Boeing.Helicopter.Team
#> 4                                    Boeing.Helicopter.Team
#> 5                                    Boeing.Helicopter.Team
#> 6                                    Boeing.Helicopter.Team
#> 7                               Four.by.Four.Rescue.Council
#> 8                               Four.by.Four.Rescue.Council
#> 9                               Four.by.Four.Rescue.Council
#> 10                              Four.by.Four.Rescue.Council
#> 11                              Four.by.Four.Rescue.Council
#> 12           German.Shepard.Search.Dogs.of.Washington.State
#> 13           German.Shepard.Search.Dogs.of.Washington.State
#> 14           German.Shepard.Search.Dogs.of.Washington.State
#> 15           German.Shepard.Search.Dogs.of.Washington.State
#> 16                                       King.County.Police
#> 17                                       King.County.Police
#> 18                                       King.County.Police
#> 19                                       King.County.Police
#> 20                                       King.County.Police
#> 21                                       King.County.Police
#> 22                Seattle.ESAR..Explorer.Search.and.Rescue.
#> 23                Seattle.ESAR..Explorer.Search.and.Rescue.
#> 24                Seattle.ESAR..Explorer.Search.and.Rescue.
#> 25                Seattle.ESAR..Explorer.Search.and.Rescue.
#> 26                Seattle.ESAR..Explorer.Search.and.Rescue.
#> 27              Washingotn.Department.of.Emergency.Services
#> 28 Washington.State.National.Guard..181st.Support.Batalion.
#> 29 Washington.State.National.Guard..181st.Support.Batalion.
#> 30 Washington.State.National.Guard..181st.Support.Batalion.
#> 31 Washington.State.National.Guard..181st.Support.Batalion.
#> 32 Washington.State.National.Guard..181st.Support.Batalion.
#> 33 Washington.State.National.Guard..181st.Support.Batalion.
#>                                                          to edge.weight
#> 1                               Four.by.Four.Rescue.Council           1
#> 2                                        King.County.Police           1
#> 3            Boeing.Employees.Amateur.Radio.Society..BEARS.           1
#> 4  Washington.State.National.Guard..181st.Support.Batalion.           1
#> 5                 Seattle.ESAR..Explorer.Search.and.Rescue.           1
#> 6                               Four.by.Four.Rescue.Council           1
#> 7                 Seattle.ESAR..Explorer.Search.and.Rescue.           1
#> 8                                        King.County.Police           1
#> 9                                    Boeing.Helicopter.Team           1
#> 10 Washington.State.National.Guard..181st.Support.Batalion.           1
#> 11           Boeing.Employees.Amateur.Radio.Society..BEARS.           1
#> 12                                       King.County.Police           1
#> 13                Seattle.ESAR..Explorer.Search.and.Rescue.           1
#> 14                              Four.by.Four.Rescue.Council           1
#> 15           Boeing.Employees.Amateur.Radio.Society..BEARS.           1
#> 16           German.Shepard.Search.Dogs.of.Washington.State           1
#> 17                Seattle.ESAR..Explorer.Search.and.Rescue.           1
#> 18                              Four.by.Four.Rescue.Council           1
#> 19              Washingotn.Department.of.Emergency.Services           1
#> 20 Washington.State.National.Guard..181st.Support.Batalion.           1
#> 21                                   Boeing.Helicopter.Team           1
#> 22                                       King.County.Police           1
#> 23                                   Boeing.Helicopter.Team           1
#> 24           German.Shepard.Search.Dogs.of.Washington.State           1
#> 25                              Four.by.Four.Rescue.Council           1
#> 26           Boeing.Employees.Amateur.Radio.Society..BEARS.           1
#> 27                                       King.County.Police           1
#> 28           German.Shepard.Search.Dogs.of.Washington.State           1
#> 29                                       King.County.Police           1
#> 30                Seattle.ESAR..Explorer.Search.and.Rescue.           1
#> 31                                   Boeing.Helicopter.Team           1
#> 32                              Four.by.Four.Rescue.Council           1
#> 33           Boeing.Employees.Amateur.Radio.Society..BEARS.           1