Function for converting a network object into the correct format for use with geomnet

Function for converting a network object into the correct format for use with geomnet

# S3 method for network
fortify(model, data = NULL, group = NULL, ...)

Arguments

model
object of class "network"

data
NULL - not used in this function

group
character. Used for facetting. If you wish to facet on a network variable provide the name of that variable here.

...
not used in this function

Examples

# class network (sna, network, statnet packages)

library(network)

#> network: Classes for Relational Data
#> Version 1.16.1 created on 2020-10-06.
#> copyright (c) 2005, Carter T. Butts, University of California-Irvine
#>                     Mark S. Handcock, University of California -- Los Angeles
#>                     David R. Hunter, Penn State University
#>                     Martina Morris, University of Washington
#>                     Skye Bender-deMoll, University of Washington
#>  For citation information, type citation("network").
#>  Type help("network-package") to get started.

#> 
#> Attaching package: ‘network’

#> The following objects are masked from ‘package:igraph’:
#> 
#>     %c%, %s%, add.edges, add.vertices, delete.edges, delete.vertices,
#>     get.edge.attribute, get.edges, get.vertex.attribute, is.bipartite,
#>     is.directed, list.edge.attributes, list.vertex.attributes,
#>     set.edge.attribute, set.vertex.attribute

data(emon, package = "network")
fortify(emon$Cheyenne)

#>    from to Frequency  na.x Command.Rank.Score Decision.Rank.Score Formalization
#> 1     1  2         1 FALSE                  0                  20             2
#> 2     1  3         1 FALSE                  0                  20             2
#> 3     1  4         1 FALSE                  0                  20             2
#> 4     1  5         1 FALSE                  0                  20             2
#> 5     1  7         1 FALSE                  0                  20             2
#> 6     1  8         1 FALSE                  0                  20             2
#> 7     1  9         3 FALSE                  0                  20             2
#> 8     1 10         1 FALSE                  0                  20             2
#> 9     1 11         1 FALSE                  0                  20             2
#> 10    1 14         1 FALSE                  0                  20             2
#> 11    2  1         1 FALSE                 10                   7             1
#> 12    2  3         3 FALSE                 10                   7             1
#> 13    2  7         1 FALSE                 10                   7             1
#> 14    2  8         4 FALSE                 10                   7             1
#> 15    2  9         2 FALSE                 10                   7             1
#> 16    2 10         2 FALSE                 10                   7             1
#> 17    2 14         3 FALSE                 10                   7             1
#> 18    3  1         2 FALSE                  3                   0             1
#> 19    3  2         1 FALSE                  3                   0             1
#> 20    3  4         4 FALSE                  3                   0             1
#> 21    3  5         3 FALSE                  3                   0             1
#> 22    3  6         3 FALSE                  3                   0             1
#> 23    3  7         4 FALSE                  3                   0             1
#> 24    3  8         3 FALSE                  3                   0             1
#> 25    3  9         1 FALSE                  3                   0             1
#> 26    3 10         1 FALSE                  3                   0             1
#> 27    3 11         1 FALSE                  3                   0             1
#> 28    3 12         1 FALSE                  3                   0             1
#> 29    3 13         3 FALSE                  3                   0             1
#> 30    4  2         4 FALSE                  5                   5             1
#> 31    4  3         2 FALSE                  5                   5             1
#> 32    4  7         1 FALSE                  5                   5             1
#> 33    4  9         1 FALSE                  5                   5             1
#> 34    4 10         1 FALSE                  5                   5             1
#> 35    5  7         4 FALSE                  0                   0             1
#> 36    5  8         1 FALSE                  0                   0             1
#> 37    5 13         1 FALSE                  0                   0             1
#> 38    6 NA        NA    NA                  0                   0             1
#> 39    7  3         1 FALSE                 20                  20             1
#> 40    7  8         1 FALSE                 20                  20             1
#> 41    7 10         1 FALSE                 20                  20             1
#> 42    7 11         2 FALSE                 20                  20             1
#> 43    7 12         1 FALSE                 20                  20             1
#> 44    7 13         1 FALSE                 20                  20             1
#> 45    8  1         1 FALSE                 40                  50             2
#> 46    8  2         1 FALSE                 40                  50             2
#> 47    8  4         1 FALSE                 40                  50             2
#> 48    8  5         1 FALSE                 40                  50             2
#> 49    8  6         3 FALSE                 40                  50             2
#> 50    8  7         3 FALSE                 40                  50             2
#> 51    8  9         1 FALSE                 40                  50             2
#> 52    8 10         1 FALSE                 40                  50             2
#> 53    8 11         1 FALSE                 40                  50             2
#> 54    8 12         4 FALSE                 40                  50             2
#> 55    8 13         1 FALSE                 40                  50             2
#> 56    8 14         1 FALSE                 40                  50             2
#> 57    9  1         1 FALSE                 10                  10             1
#> 58    9  6         1 FALSE                 10                  10             1
#> 59    9  7         3 FALSE                 10                  10             1
#> 60    9  8         1 FALSE                 10                  10             1
#> 61    9 10         1 FALSE                 10                  10             1
#> 62    9 13         1 FALSE                 10                  10             1
#> 63    9 14         1 FALSE                 10                  10             1
#> 64   10  7         1 FALSE                 30                  20             3
#> 65   10  9         1 FALSE                 30                  20             3
#> 66   11  7         2 FALSE                  2                   3            NA
#> 67   11  9         2 FALSE                  2                   3            NA
#> 68   11 10         3 FALSE                  2                   3            NA
#> 69   11 13         1 FALSE                  2                   3            NA
#> 70   11 14         2 FALSE                  2                   3            NA
#> 71   12  3         2 FALSE                  2                   3             1
#> 72   12  7         3 FALSE                  2                   3             1
#> 73   12  8         1 FALSE                  2                   3             1
#> 74   12  9         1 FALSE                  2                   3             1
#> 75   13  3         4 FALSE                  2                   2             1
#> 76   13  7         3 FALSE                  2                   2             1
#> 77   13  8         3 FALSE                  2                   2             1
#> 78   13  9         1 FALSE                  2                   2             1
#> 79   13 10         3 FALSE                  2                   2             1
#> 80   13 11         1 FALSE                  2                   2             1
#> 81   13 12         1 FALSE                  2                   2             1
#> 82   14  1         1 FALSE                  0                   0             1
#> 83   14  5         1 FALSE                  0                   0             1
#> 84   14  8         1 FALSE                  0                   0             1
#>    Location  na.y Paid.Staff Sponsorship
#> 1         L FALSE         10       State
#> 2         L FALSE         10       State
#> 3         L FALSE         10       State
#> 4         L FALSE         10       State
#> 5         L FALSE         10       State
#> 6         L FALSE         10       State
#> 7         L FALSE         10       State
#> 8         L FALSE         10       State
#> 9         L FALSE         10       State
#> 10        L FALSE         10       State
#> 11        L FALSE        400       State
#> 12        L FALSE        400       State
#> 13        L FALSE        400       State
#> 14        L FALSE        400       State
#> 15        L FALSE        400       State
#> 16        L FALSE        400       State
#> 17        L FALSE        400       State
#> 18        L FALSE        200       State
#> 19        L FALSE        200       State
#> 20        L FALSE        200       State
#> 21        L FALSE        200       State
#> 22        L FALSE        200       State
#> 23        L FALSE        200       State
#> 24        L FALSE        200       State
#> 25        L FALSE        200       State
#> 26        L FALSE        200       State
#> 27        L FALSE        200       State
#> 28        L FALSE        200       State
#> 29        L FALSE        200       State
#> 30        L FALSE         60     Federal
#> 31        L FALSE         60     Federal
#> 32        L FALSE         60     Federal
#> 33        L FALSE         60     Federal
#> 34        L FALSE         60     Federal
#> 35        L FALSE          1     Private
#> 36        L FALSE          1     Private
#> 37        L FALSE          1     Private
#> 38        L FALSE          7     Private
#> 39        L FALSE         60      County
#> 40        L FALSE         60      County
#> 41        L FALSE         60      County
#> 42        L FALSE         60      County
#> 43        L FALSE         60      County
#> 44        L FALSE         60      County
#> 45        L FALSE          7 County/City
#> 46        L FALSE          7 County/City
#> 47        L FALSE          7 County/City
#> 48        L FALSE          7 County/City
#> 49        L FALSE          7 County/City
#> 50        L FALSE          7 County/City
#> 51        L FALSE          7 County/City
#> 52        L FALSE          7 County/City
#> 53        L FALSE          7 County/City
#> 54        L FALSE          7 County/City
#> 55        L FALSE          7 County/City
#> 56        L FALSE          7 County/City
#> 57        L FALSE         70        City
#> 58        L FALSE         70        City
#> 59        L FALSE         70        City
#> 60        L FALSE         70        City
#> 61        L FALSE         70        City
#> 62        L FALSE         70        City
#> 63        L FALSE         70        City
#> 64        L FALSE        100        City
#> 65        L FALSE        100        City
#> 66        L FALSE         NA      County
#> 67        L FALSE         NA      County
#> 68        L FALSE         NA      County
#> 69        L FALSE         NA      County
#> 70        L FALSE         NA      County
#> 71        L FALSE          1      County
#> 72        L FALSE          1      County
#> 73        L FALSE          1      County
#> 74        L FALSE          1      County
#> 75        L FALSE         10     Private
#> 76        L FALSE         10     Private
#> 77        L FALSE         10     Private
#> 78        L FALSE         10     Private
#> 79        L FALSE         10     Private
#> 80        L FALSE         10     Private
#> 81        L FALSE         10     Private
#> 82        L FALSE          0     Private
#> 83        L FALSE          0     Private
#> 84        L FALSE          0     Private
#>                                      vertex.names Volunteer.Staff
#> 1       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 2       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 3       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 4       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 5       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 6       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 7       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 8       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 9       Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 10      Wyoming.Disaster.and.Civil.Defense.Agnecy              50
#> 11    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 12    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 13    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 14    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 15    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 16    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 17    Wyoming.State.National.Guard..Army.and.Air.            2000
#> 18                   Wyoming.State.Highway.Patrol               0
#> 19                   Wyoming.State.Highway.Patrol               0
#> 20                   Wyoming.State.Highway.Patrol               0
#> 21                   Wyoming.State.Highway.Patrol               0
#> 22                   Wyoming.State.Highway.Patrol               0
#> 23                   Wyoming.State.Highway.Patrol               0
#> 24                   Wyoming.State.Highway.Patrol               0
#> 25                   Wyoming.State.Highway.Patrol               0
#> 26                   Wyoming.State.Highway.Patrol               0
#> 27                   Wyoming.State.Highway.Patrol               0
#> 28                   Wyoming.State.Highway.Patrol               0
#> 29                   Wyoming.State.Highway.Patrol               0
#> 30   Francis.E..Warren.Air.Force.Base..fire.unit.               0
#> 31   Francis.E..Warren.Air.Force.Base..fire.unit.               0
#> 32   Francis.E..Warren.Air.Force.Base..fire.unit.               0
#> 33   Francis.E..Warren.Air.Force.Base..fire.unit.               0
#> 34   Francis.E..Warren.Air.Force.Base..fire.unit.               0
#> 35                                      Red.Cross              20
#> 36                                      Red.Cross              20
#> 37                                      Red.Cross              20
#> 38                                 Salvation.Army              80
#> 39                Laramie.County.Sheriff.s.Office              20
#> 40                Laramie.County.Sheriff.s.Office              20
#> 41                Laramie.County.Sheriff.s.Office              20
#> 42                Laramie.County.Sheriff.s.Office              20
#> 43                Laramie.County.Sheriff.s.Office              20
#> 44                Laramie.County.Sheriff.s.Office              20
#> 45 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 46 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 47 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 48 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 49 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 50 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 51 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 52 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 53 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 54 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 55 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 56 Laramie.County...Cheyenne.Civil.Defense.Agency             100
#> 57                       Cheyenne.Fire.Department               0
#> 58                       Cheyenne.Fire.Department               0
#> 59                       Cheyenne.Fire.Department               0
#> 60                       Cheyenne.Fire.Department               0
#> 61                       Cheyenne.Fire.Department               0
#> 62                       Cheyenne.Fire.Department               0
#> 63                       Cheyenne.Fire.Department               0
#> 64                     Cheyenne.Police.Department               0
#> 65                     Cheyenne.Police.Department               0
#> 66    Laramie.County.Fire.District..1..volunteer.              NA
#> 67    Laramie.County.Fire.District..1..volunteer.              NA
#> 68    Laramie.County.Fire.District..1..volunteer.              NA
#> 69    Laramie.County.Fire.District..1..volunteer.              NA
#> 70    Laramie.County.Fire.District..1..volunteer.              NA
#> 71    Laramie.County.Fire.District..2..volunteer.              20
#> 72    Laramie.County.Fire.District..2..volunteer.              20
#> 73    Laramie.County.Fire.District..2..volunteer.              20
#> 74    Laramie.County.Fire.District..2..volunteer.              20
#> 75                          A.1.Ambulance.Service               6
#> 76                          A.1.Ambulance.Service               6
#> 77                          A.1.Ambulance.Service               6
#> 78                          A.1.Ambulance.Service               6
#> 79                          A.1.Ambulance.Service               6
#> 80                          A.1.Ambulance.Service               6
#> 81                          A.1.Ambulance.Service               6
#> 82                          Shy.Wy.HAM.Radio.Club              NA
#> 83                          Shy.Wy.HAM.Radio.Club              NA
#> 84                          Shy.Wy.HAM.Radio.Club              NA